[Decomp][Player] Cohesion split: extract pet/charm spell init to PlayerPet.cpp
